+++ src/gwt/src/org/rstudio/studio/client/workbench/views/source/editors/text/DocDisplay.java
@@ -263,6 +263,7 @@ public interface DocDisplay extends HasValueChangeHandlers<Void>,
    void removeMarkersOnCursorLine();
    
    void beginCollabSession(String serverUrl);
+   void endCollabSession();
 
    void setPopupVisible(boolean visible);
    boolean isPopupVisible();